segformer_segmentation/requirements.txt
torch>=2.0
torchvision>=0.15
transformers>=4.36
opencv-python
numpy
matplotlib
Pillow
scikit-learn
